# projects/apps/divergesapp/backend/openapi.yaml
openapi: 3.0.0
info:
  title: DivergesiApp API
  version: 1.0.0
  description: API for DivergesiApp, an interactive educational platform
servers:
  - url: http://localhost:1337
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        uid:
          type: string
        email:
          type: string
        full_name:
          type: string
        user_type:
          type: string
          enum:
            - teacher
            - student
            - guardian
        created_at:
          type: integer
        updated_at:
          type: integer
    UserProfile:
      type: object
      properties:
        uid:
          type: string
        user_uid:
          type: string
        profile_picture_url:
          type: string
        bio:
          type: string
        grade_level:
          type: integer
        subjects_taught:
          type: object
        contact_information:
          type: object
    Course:
      type: object
      properties:
        uid:
          type: string
        teacher_uid:
          type: string
        name:
          type: string
        description:
          type: string
        grade_level:
          type: integer
        subject:
          type: string
        created_at:
          type: integer
        updated_at:
          type: integer
    Module:
      type: object
      properties:
        uid:
          type: string
        course_uid:
          type: string
        name:
          type: string
        description:
          type: string
        order_index:
          type: integer
        prerequisites:
          type: object
...
paths:
  /auth/login:
    post:
      summary: Authenticate user and get token
      operationId: login
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
      responses:
        "200":
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
  /auth/signup:
    post:
      summary: Register a new user
      operationId: signup
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                full_name:
                  type: string
                user_type:
                  type: string
                  enum:
                    - teacher
                    - student
                    - guardian
              required:
                - email
                - password
                - full_name
                - user_type
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
  /users/{user_uid}/profile:
    get:
      summary: Get user profile
      operationId: get_user_profile
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_uid
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      summary: Update user profile
      operationId: update_user_profile
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_uid
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfile"
      responses:
        "200":
          description: User profile updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /courses:
    post:
      summary: Create a new course
      operationId: create_course
      tags:
        - Courses
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                grade_level:
                  type: integer
                subject:
                  type: string
              required:
                - name
                - grade_level
                - subject
      responses:
        "201":
          description: Course created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    get:
      summary: Get all courses
      operationId: get_all_courses
      tags:
        - Courses
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Courses retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
  /courses/{course_uid}:
    get:
      summary: Get course details
      operationId: get_course
      tags:
        - Courses
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: course_uid
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Course details retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    put:
      summary: Update course details
      operationId: update_course
      tags:
        - Courses
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: course_uid
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Course"
      responses:
        "200":
          description: Course updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      summary: Delete a course
      operationId: delete_course
      tags:
        - Courses
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: course_uid
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Course deleted successfully
  /courses/{course_uid}/modules:
    post:
      summary: Create a new module in a course
      operationId: create_module
      tags:
        - Modules
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: course_uid
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Module"
      responses:
        "201":
          description: Module created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
    get:
      summary: Get all modules in a course
      operationId: get_course_modules
      tags:
        - Modules
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: course_uid
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Modules retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Module"
  /modules/{module_uid}/content:
    post:
      summary: Add content to a module
      operationId: add_module_content
      tags:
        - Content
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: module_uid
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                content_type:
                  type: string
                  enum:
                    - video
                    - pdf
                    - image
                    - text
                    - presentation
                    - audio
                is_mandatory:
                  type: boolean
                file:
                  type: string
                  format: binary
              required:
                - title
                - content_type
                - file
      responses:
        "201":
          description: Content added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
    get:
      summary: Get all content in a module
      operationId: get_module_content
      tags:
        - Content
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: module_uid
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Content retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentItem"
  /courses/{course_uid}/assignments:
    post:
      summary: Create a new assignment
      operationId: create_assignment
      tags:
        - Assignments
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: course_uid
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Assignment"
      responses:
        "201":
          description: Assignment created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
    get:
      summary: Get all assignments in a course
      operationId: get_course_assignments
      tags:
        - Assignments
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: course_uid
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Assignments retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
  /assignments/{assignment_uid}/submit:
    post:
      summary: Submit an assignment
      operationId: submit_assignment
      tags:
        - Assignments
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: assignment_uid
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                submission_text:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Assignment submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
...
security:
  - bearerAuth: []